backend/requirements.txt
# Delege Backend Dependencies
# Python 3.11+ required

# Web Framework
fastapi>=0.115.0
uvicorn[standard]>=0.32.0

# Agent Framework
agno>=1.5.0

# Task Queue
celery>=5.4.0
redis>=5.2.0

# Database (pinned to avoid pyiceberg dependency)
supabase>=2.9.0,<2.10.0

# LLM (unified SDK, required by Agno's Gemini integration as well)
google-genai>=1.0.0

# Document Processing (for knowledge base)
pypdf>=5.0.0
python-docx>=1.1.0
python-multipart>=0.0.18

# Text Chunking & Embeddings
tiktoken>=0.8.0

# Environment Variables
python-dotenv>=1.0.1

# HTTP Client
httpx>=0.27.0,<0.28.0

# Utilities
pydantic>=2.10.0
pydantic-settings>=2.6.0
